# Optimized Implementation of `Array#fill` for Zero Values