See discussion here:

https://forum.crystal-lang.org/t/fastest-and-iconic-way-to-zero-an-array/1799/16

This PR provides a high performance (~10x) means to initialize arrays of integers to zero by optimizing that case in `Array#fill`.

```
Using compiled compiler at .build/crystal
          fill(0)   2.50M (400.01ns) (± 1.82%)  0.0B/op   1.43× slower
       fill(0, 0)   2.52M (397.50ns) (± 2.39%)  0.0B/op   1.42× slower
fill(0, 0, 1_000)   3.57M (279.82ns) (± 2.95%)  0.0B/op        fastest
      fill(0, ..)   2.48M (403.21ns) (± 3.81%)  0.0B/op   1.44× slower
```
```
Using compiled compiler at .build/crystal
          fill(0)  23.72M ( 42.16ns) (± 4.14%)  0.0B/op   1.15× slower
       fill(0, 0)  25.42M ( 39.34ns) (± 2.81%)  0.0B/op   1.08× slower
fill(0, 0, 1_000)  27.29M ( 36.65ns) (± 5.10%)  0.0B/op   1.00× slower
      fill(0, ..)  27.37M ( 36.54ns) (± 3.41%)  0.0B/op        fastest
```