@asterite 2020-03-12
Thank you for this! ❤️ 

> Using compiled compiler at .build/crystal

Just a note: when making changes to the standard library (not the compiler) there's no need to compiler the compiler to try things out. Just using `bin/crystal` will pick the new standard library (while `crystal` will use the installed standard library).