@toddsundsted 2020-03-13
6769781 adds support for arrays for floating point numbers, as well. this should cover arrays of all non-union numeric types, and should be suitable for most computational purposes.